- Seperate drawing of a room into its own component for dynamic rendering and addition
